Rework zoom implementation
This commit rewrites the entire implementation of the previous commit
from scratch. The camera is now represented as a rectangle with a
variable width and height, and it moves around the statically sized map.
The map is still resized and shifted around under the hood, but this
approach is much more intuitive and causes far less complexity overall.